+++ src/NodeJsServer.js
@@ -90,7 +90,7 @@ module.exports = class NodeJsServer extends DockerServer {
         } catch (error) {
         }
 
-        await fs.rename(serverDir, serverDir + "-old");
+        if (existsSync(serverDir)) await fs.rename(serverDir, serverDir + "-old");
         await fs.rename(tempDir, serverDir);
 
         this.lastLogs = [];
