+++ src/PythonServer.js
@@ -74,7 +74,7 @@ module.exports = class PythonServer extends DockerServer {
         } catch (error) {
         }
 
-        await fs.rename(serverDir, serverDir + "-old");
+        if (existsSync(serverDir)) await fs.rename(serverDir, serverDir + "-old");
         await fs.rename(tempDir, serverDir);
 
         this.lastLogs = [];
